custom_modules/requirements.txt
opencv-contrib-python
numpy

pytesseract

pyzbar
qrcode
python-barcode

scikit-learn
imutils
